Sprint 1 - Frontend cleanup (#32)
