+++ README.md
@@ -119,167 +119,6 @@ Automated user interaction tests are executed on pull requests.  The test is rec
 ![](common/graphics/cypress-artifacts.png)
 
 
-# Starter Application
-
-The starter stack includes a frontend, backend and postgres database.  The frontend and backend are buld with [NestJS](https://docs.nestjs.com).  They currently do very little, but provide placeholders for more functional products.  See subfolder for source, including Dockerfiles and OpenShift templates.
-
-Features:
-* [TypeScript](https://www.typescriptlang.org/) strong-typing for JavaScript
-* [NestJS](https://docs.nestjs.com) frontend and backend
-* [Postgres](https://www.postgresql.org/) or [PostGIS](https://postgis.net/) database
-* [backup-container](https://github.com/BCDevOps/backup-container) provided by BCDevOps
-
-Postgres is default.  Switch to PostGIS by copying the appropriate Dockerfile to `./database`:
-
-> cp ./database/postgis/Dockerfile ./database
-
-Test and develop locally with Docker Compose:
-
-> docker-compose up -d
-
-**Example APIs, UIs and Metabase/Oracle Templates**
-
-Templates for APIs, UIs and Metabase/Oracle can be used to kickstart or extend projects.  Please visit our collaborators' [NR Architecture Templates](https://github.com/bcgov/nr-arch-templates) repository for more information.
-
-# Getting Started
-
-Initial setup is intended to take four hours or less.  This depends greatly on intended complexity, features selected/excluded and outside cooperation.
-
-## Prerequisites
-
-The following are required:
-
-* BC Government IDIR accounts for anyone submitting requests
-* GitHub accounts for all participating team members
-    * [Sign Up is free](https://github.com/signup)
-* Membership in the BCGov GitHub organization
-    * Provide GitHub IDs to [BCGov's Just Ask](https://just-ask.developer.gov.bc.ca/)
-* Project namespaces:
-    * OpenShift - [Register a New Project](https://registry.developer.gov.bc.ca/public-landing)
-
-
-
-## GitHub Secrets, Variables and Environments
-
-Variables and secrets are consumed by workflows.  Environments provide their own sets of secrets and variables, overriding default sets.
-
-### Repository Secrets and Variables
-
-Repository secrets and variables are available to all workflows, except pull requests triggered by Dependabot.
-
-Secrets are hidden from logs and outputs, while variables are visible.  Using secrets exclusively can make troubeshooting more difficult.
-
-> Click Settings > Secrets and Variables > Actions > Secrets > New repository secret
-
-> Click Settings > Secrets and Variables > Actions > Variables > New repository variable
-
-### Environment
-
-Environments are groups of secrets and variables that can be gatekept.  This includes limting access to certain users or requiring manual approval before a requesting workflow can run.  Environment values override any default values.
-
-> Click Settings > Environments > New environment
-
-Environments provide a [number of features](https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment), including:
-
-* Required reviewers
-* Wait timer
-* Deployment branches
-
-## Secret and Variable Values
-
-### Secrets
-
-**GITHUB_TOKEN**
-
-Default token.  Replaced every workflow run, available to all workflows.
-* Consume: `{{ secrets.GITHUB_TOKEN }}`
-
-**OC_TOKEN**
-
-OpenShift token, different for every project/namespace.  This guide assumes your OpenShift platform team has provisioned a pipeline account.
-
-* Consume: `{{ secrets.OC_TOKEN }}`
-
-Locate an OpenShift pipeline token:
-
-1. Login to your OpenShift cluster, e.g.: [Gold](https://console.apps.silver.devops.gov.bc.ca/) or [Silver](https://console.apps.silver.devops.gov.bc.ca/)
-2. Select your DEV namespace
-3. Click Workloads > Secrets (under Workloads for Administrator view)
-4. Select `pipeline-token-...` or a similarly privileged token
-5. Under Data, copy `token`
-6. Paste into the GitHub Secret `OC_TOKEN` (see above)
-
-**SONAR_TOKEN and Other Sonar Tokens**
-
-If SonarCloud is being used each application will have its own token.  Single-application repositories typically use `${{ secrets.SONAR_TOKEN }}`, but monoreposities will have multiple, like `${{ secrets.SONAR_TOKEN_BACKEND }}` and `${{ secrets.SONAR_TOKEN_FRONTEND }}`.
-
-BC Government employees can request SonarCloud projects from [bcdevops/devops-requests](https://github.com/BCDevOps/devops-requests) by creating a SonarCloud request/[issue](https://github.com/BCDevOps/devops-requests/issues/new/choose).  This template expects a monorepo, so please ask for that and provide component names (e.g. backend, frontend).
-
-### Variables
-
-**OC_SERVER**
-
-OpenShift server address.
-* Consume: `{{ vars.OC_SERVER }}`
-* Value: `https://api.silver.devops.gov.bc.ca:6443`
-
-**OC_NAMESPACE**
-
-OpenShift project/namespace.  Provided by your OpenShift platform team.
-
-* Consume: `{{ vars.OC_NAMESPACE }}`
-* Value: format `c1c7ed-dev | test | prod`
-
-## Repository Configuration
-
-### Pull Request Handling
-
-Squash merging is recommended for simplified history and ease of rollback.  Cleaning up merged branches is recommended for your DevOps Specialist's fragile sanity.
-
-> Click Settings > General (selected automatically)
-
-Pull Requests:
-
-- `[uncheck] Allow merge commits`
-- `[check] Allow squash merging`
-  - `Default to pull request title`
-- `[uncheck] Allow rebase merging`
-- `[check] Always suggest updating pull request branches`
-- `[uncheck] Allow auto-merge`
-- `[check] Automatically delete head branches`
-
-### Packages
-
-Packages are available from your repository (link on right).  All should have visibility set to public for the workflows to run successfully.
-
-https://github.com/orgs/bcgov/packages?repo_name=nr-compliance-enforcement
-
-### Branch Protection
-
-This is required to prevent direct pushes and merges to the default branch.  These steps must be run after one full pull request pipeline has been run.
-
-1. Select Settings (gear, top right) -> Branches (under Code and Automation)
-2. Click `Add Rule` or edit an existing rule
-3. Under `Protect matching branches` specify the following:
-    * Branch name pattern: `main`
-    * `[check] Require a pull request before merging`
-        * `[check] Require approvals` (default = 1)
-        * `[check] Dismiss stale pull request approvals when new commits are pushed`
-        * `[check] Require review from Code Owners`
-    * `[check] Require status checks to pass before merging`
-        * `[check] Require branches to be up to date before merging`
-        * `Status checks that are required`:
-            * Select checks as appropriate, e.g. Build x, Deploy y
-    * `[check] Require conversation resolution before merging`
-    * `[check] Include administrators` (optional)
-
-# Natural Resources Kickstarter
-
-Members of the BC Government's Natural Resource minisistries are strongly recommended to follow the recommendations in their [Kickstarter Guide](https://github.com/bcgov/nr-arch-templates/blob/main/confluence/pages/Agile_Team_Kickstarter/README.md).  The linked document is generated from Confluence, so some links may be internal-only (sorry!).
-
-[Natural Resources Kickstarter Guide](https://github.com/bcgov/nr-arch-templates/blob/main/confluence/pages/Agile_Team_Kickstarter/README.md)
-
-
 # Feedback
 
 Please contribute your ideas!  [Issues](/../../issues) and [Pull Requests](/../../pulls) are appreciated.
