+++ frontend/src/app/App.tsx
@@ -1,27 +1,16 @@
+import { BrowserRouter } from "react-router-dom";
 import { Counter } from "./features/counter/Counter";
 import "./App.css";
-import React from "react";
-import { Provider } from "react-redux";
-import { store } from "./store/store";
-import { BrowserRouter } from "react-router-dom";
-import RenderOnRole from "../components/RenderOnRole";
 import Roles from "./constants/roles";
-// import { SpaceReference as Reference } from "./components/reference";
+import RenderOnRole from "./components/routing/render-on-role";
 
 function App() {
   return (
-    <React.StrictMode>
-      <Provider store={store}>
-        <BrowserRouter>
-          <div className="container">
-            <RenderOnRole roles={[Roles.COS_OFFICER]}>
-              <Counter/>
-              {/* <Reference /> */}
-            </RenderOnRole>
-          </div>
-        </BrowserRouter>
-      </Provider>
-    </React.StrictMode>
+    <BrowserRouter>
+      <RenderOnRole roles={[Roles.COS_OFFICER]}>
+        <Counter />
+      </RenderOnRole>
+    </BrowserRouter>
   );
 }
