+++ frontend/src/app/components/routing/index.ts
@@ -1,0 +1,8 @@
+import NotAllowed from "./not-allowed";
+import RenderOnAnonymous from "./render-on-anonymous";
+import RenderOnAuthenticated from "./render-on-authenticated";
+import RenderOnRole from "./render-on-role";
+
+export { NotAllowed, RenderOnAnonymous, RenderOnAuthenticated };
+
+export default RenderOnRole;
