+++ frontend/src/app/components/routing/render-on-authenticated.jsx
@@ -1,4 +1,4 @@
-import UserService from "../app/service/UserServices";
+import UserService from "../app/service/user-service";
 
 
 const RenderOnAuthenticated = ({ children }) => (UserService.isLoggedIn()) ? children : null;
