+++ frontend/src/index.tsx
@@ -1,18 +1,28 @@
-import reportWebVitals from "./reportWebVitals";
-import App from "./app/App";
+import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import UserService from "./app/service/UserServices";
+import { Provider } from "react-redux";
+import { store } from "./app/store/store";
 
-import "./assets/sass/app.scss"
+import App from "./app/App";
+import UserService from "./app/service/user-service";
+
+import reportWebVitals from "./reportWebVitals";
 
 const container = document.getElementById("root")!;
 const root = createRoot(container);
 
-const renderApp = () => root.render(<App />);
+const onAuthenticatedCallback = () =>
+  root.render(
+    <StrictMode>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </StrictMode>
+  );
+
+UserService.initKeycloak(onAuthenticatedCallback);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-
-UserService.initKeycloak(renderApp);
